main_project/requirements.txt
streamlit
pandas
numpy
Pillow